Add files via upload
This code sample takes a string as an input and gives an encrypted output. There are 2 stages of encryption and the output at each stage is displayed. The decryption of the final output is also done and the output of each stage of decryption is also displayed in the reverse order.